feat: Abstract out the KV cache from of the repo
This commit abstracts the key value cache from the key value repository.
We do this so that we can install the kv repository as a transient as
only the kv cache has to be a singleton.

We need to do this as we want to be able to inject the database context
into the kv repository. However, as database contexts should be scoped
services, we need the repository to be transient/scoped.