feat: Create a kubernetes deployment for KVS and a LoadBalancer
We now have a deployment file for deploying our kvs image with
replicated nodes. We also have a load balancer for sending connections
to each of the KVS nodes.

Currently, we don't have any deployment files for dependencies, but they
can currently manually be created using helm.

We use the secret files rabbit-mq-credentials and postgres-credentials
for storing our secrets for RabbitMQ and postgres (resp.)